remove fly.io config
