--- .github/workflows/fly.yml
@@ -1,20 +1,0 @@
-name: Fly Deploy 🔮
-on:
-  push:
-    branches:
-      - dev
-jobs:
-  deploy:
-    name: Deploy app 🚢
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v3
-      - uses: superfly/flyctl-actions/setup-flyctl@master
-      - run: |
-          flyctl deploy --remote-only \
-          --build-secret DATABASE_URL=${{ secrets.DATABASE_URL }} \
-          --build-secret DATABASE_AUTH_TOKEN=${{ secrets.DATABASE_AUTH_TOKEN }}
-            
-        env:
-          FLY_API_TOKEN: ${{ secrets.FLY_API_TOKEN }}
-
