--- fly.toml
@@ -1,17 +1,0 @@
-# fly.toml app configuration file generated for genug-sveltekit on 2023-10-01T16:33:59+02:00
-#
-# See https://fly.io/docs/reference/configuration/ for information about how to use this file.
-#
-
-app = "genug-sveltekit"
-primary_region = "ams"
-
-[build]
-
-[http_service]
-  internal_port = 3000
-  force_https = true
-  auto_stop_machines = true
-  auto_start_machines = true
-  min_machines_running = 0
-  processes = ["app"]
